requirements.txt
rich>=13.0.0
requests>=2.31.0
websocket-client>=1.7.0
# Future additions (uncomment when backend supports them):
# websockets>=12.0          # Async WebSocket alternative